LayoutTests/fast/js/date-DST-time-cusps-expected.txt

For times that happen twice the behavior of all major browsers seems to be to pick the second occurance, i.e. Standard Time not Daylight Time

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new Date(1982, 2, 14, 2, 10)).getHours() is 1
PASS (new Date(1982, 2, 14, 2)).getHours() is 1
PASS (new Date(1982, 11, 7, 1, 10)).getTimezoneOffset() is 480
PASS (new Date(1982, 11, 7, 1)).getTimezoneOffset() is 480
PASS successfullyParsed is true

TEST COMPLETE
